fix: Improve test TLS domain assignment logic
Specifically, add the minute of the hour to the day of the year and
mod that by 12 to get the domain number (with 0 being replaced by
12). This will give us a different domain every twelve minutes, with
the starting point of the rotation changing every day so that we don't
end up using the same domain for every cronjob run.

This way, we avoid rate limit issues by using the same domain for the
same window each day, as we do currently, and we get a different
domain more frequently, decreasing the chances of two PRs stomping on
each other.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>